bugfix thread log inside baseSetting is not working xd
