+++ config.yaml
@@ -17,7 +17,7 @@ dispatcher:
 
 metrics:
   group_id: "metrics"
-  host: "0.0.0.0"
+  host: "127.0.0.1"
   port: 8000
 
 drivers_loader:
