--- shared/config/clock.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class ClockConfig(BaseModel):
-    advance_delta_seconds: int = 1
