--- shared/config/drivers_loader.py
@@ -1,7 +1,0 @@
-from pathlib import Path
-
-from pydantic import BaseModel
-
-
-class DriverLoaderConfig(BaseModel):
-    json_file: Path
