+++ shared/files.py
@@ -5,7 +5,7 @@ import ijson
 from pydantic import BaseModel
 
 
-def load_models_from_json(filepath: Path, ijson_key: str, model_cls: BaseModel) -> Iterator[BaseModel]:
+def load_models_from_json(filepath: Path, ijson_key: str, model_cls: type[BaseModel]) -> Iterator[BaseModel]:
     """
     Yield T objects one by one from a JSON file.
     Expects JSON structure: { "<ijson_key>": [ {...}, {...} ] } where {...} Structure = T
